Removed `Requests` object
